Corrections to script to match N-Central v11
The API was changed (or changed to match the documentation) making the calls in the script to get the list of customers and sites fail.  The script has been updated to use the currently documented call.  Rename V11-InstallFromNCentralAgent.aspx.cs to InstallFromNCentralAgent.aspx.cs in your deployment if using N-Central v11.